+++ sentry-samples/sentry-samples-spring-jakarta/src/main/java/io/sentry/samples/spring/jakarta/web/MetricController.java
@@ -24,7 +24,8 @@ public class MetricController {
 
   @GetMapping("gauge/{value}")
   String gauge(@PathVariable("value") Long value) {
-    Sentry.metrics().gauge("memory.free", value.doubleValue(), MeasurementUnit.Information.BYTE);
+    Sentry.metrics()
+        .gauge("memory.free", value.doubleValue(), MeasurementUnit.Information.BYTE.apiName());
     return "gauge metric tracked";
   }
 
@@ -32,7 +33,9 @@ public class MetricController {
   String distribution(@PathVariable("value") Long value) {
     Sentry.metrics()
         .distribution(
-            "distributionMetric", value.doubleValue(), MeasurementUnit.Duration.MILLISECOND);
+            "distributionMetric",
+            value.doubleValue(),
+            MeasurementUnit.Duration.MILLISECOND.apiName());
     return "distribution metric tracked";
   }
 }