+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Internal
 
-- Deprecate `MetricsUnit` in favor of `MeasurementUnit`, which defines the same units and is now accepted directly by the `Sentry.metrics()` API ([#6086](https://github.com/getsentry/sentry-java/pull/6086))
+- Deprecate `MetricsUnit` in favor of `MeasurementUnit`, which defines the same units; pass them to the `Sentry.metrics()` API as `MeasurementUnit.Duration.MILLISECOND.apiName()` ([#6086](https://github.com/getsentry/sentry-java/pull/6086))
 
 ## 8.56.0